+++ pyscript/core/cache.py
@@ -1,15 +1,54 @@
 from .bases import Pys
-from .constants import LIBRARIES_PATH, SITE_PACKAGES_PATH
-from .exceptions import PysTraceback
-from .utils.debug import print_display, print_traceback
+from .constants import PYSCRIPT_PATH, CORE_PATH, LIBRARIES_PATH, SITE_PACKAGES_PATH
+from .utils.debug import print_display, print_traceback, clear_shell
 from .utils.decorators import inheritable, singleton
 
-from typing import Any, Callable, Literal
+from typing import Literal
+from types import ModuleType
 
-loading_modules = set()
-path = [SITE_PACKAGES_PATH, LIBRARIES_PATH]
-modules = {}
-singletons = {}
+import sys
+
+pys_sys = ModuleType(
+    'sys',
+    "This is a sys module for PyScript, which provides necessary attributes, functions, and objects for the PyScript "
+    "environment. It is designed to be used within the PyScript runtime and should not be confused with the standard "
+    "Python sys module."
+)
+
+pys_sys.__running_shell__ = False
+pys_sys.__running_breakpoint__ = False
+pys_sys.argv = ['']
+pys_sys.displayhook = print_display
+pys_sys.excepthook = print_traceback
+pys_sys.clearhook = clear_shell
+pys_sys.exec_prefix = pys_sys.prefix = pys_sys.pyscript_path = PYSCRIPT_PATH
+pys_sys.core_path = CORE_PATH
+pys_sys.libraries_path = LIBRARIES_PATH
+pys_sys.site_packages_path = SITE_PACKAGES_PATH
+pys_sys.executable = f'"{sys.executable}" -m pyscript'
+pys_sys.ps1 = '>>> '
+pys_sys.ps2 = '... '
+pys_sys.float_info = sys.float_info
+pys_sys.int_info = sys.int_info
+pys_sys.hash_info = sys.hash_info
+pys_sys.stderr = sys.stderr
+pys_sys.stdout = sys.stdout
+pys_sys.stdin = sys.stdin
+pys_sys.platform = sys.platform
+pys_sys.exit = sys.exit
+pys_sys.getrecursionlimit = sys.getrecursionlimit
+pys_sys.getrefcount = sys.getrefcount
+pys_sys.getsizeof = sys.getsizeof
+pys_sys.setrecursionlimit = sys.setrecursionlimit
+
+# added in python>=3.10.7
+if hasattr(sys, 'set_int_max_str_digits'):
+    pys_sys.set_int_max_str_digits = sys.set_int_max_str_digits
+
+pys_sys.path = [SITE_PACKAGES_PATH, LIBRARIES_PATH]
+pys_sys.loading_modules = set()
+pys_sys.modules = {}
+pys_sys.singletons = {}
 
 @singleton
 @inheritable
@@ -28,92 +67,4 @@ class PysUndefined(Pys):
     def __bool__(self) -> Literal[False]:
         return False
 
-@singleton
-@inheritable
-class PysHook(Pys):
-
-    __slots__ = ()
-
-    def __new_singleton__(cls) -> 'PysHook':
-        global hook
-        hook = super(cls, cls).__new__(cls)
-        hook.argv = ['']
-        hook.running_shell = False
-        hook.running_breakpoint = False
-        hook.display = print_display
-        hook.exception = print_traceback
-        hook.ps1 = '>>> '
-        hook.ps2 = '... '
-        return hook
-
-    def __repr__(self) -> str:
-        return f'<hook object at {id(self):016X}>'
-
-    @property
-    def argv(self) -> list[str]:
-        return singletons['hook.argv']
-
-    @argv.setter
-    def argv(self, value: list[str]) -> None:
-        if not isinstance(value, list) or not all(isinstance(arg, str) for arg in value):
-            raise TypeError("hook.argv: argv must be list of strings")
-        singletons['hook.argv'] = value
-
-    @property
-    def running_shell(self) -> bool:
-        return singletons['hook.running_shell']
-
-    @running_shell.setter
-    def running_shell(self, value: bool) -> None:
-        singletons['hook.running_shell'] = bool(value)
-
-    @property
-    def running_breakpoint(self) -> bool:
-        return singletons['hook.running_breakpoint']
-
-    @running_breakpoint.setter
-    def running_breakpoint(self, value: bool) -> None:
-        singletons['hook.running_breakpoint'] = bool(value)
-
-    @property
-    def display(self) -> Callable[[Any], None]:
-        return singletons['hook.display']
-
-    @display.setter
-    def display(self, value: Callable[[Any], None]) -> None:
-        if value is not None and not callable(value):
-            raise TypeError("hook.display: must be callable")
-        singletons['hook.display'] = value
-
-    @property
-    def exception(self) -> Callable[[type[BaseException], BaseException | None, PysTraceback], None]:
-        return singletons['hook.exception']
-
-    @exception.setter
-    def exception(self, value: Callable[[type[BaseException], BaseException | None, PysTraceback], None]) -> None:
-        if value is not None and not callable(value):
-            raise TypeError("hook.exception: must be callable")
-        singletons['hook.exception'] = value
-
-    @property
-    def ps1(self) -> str:
-        return singletons['hook.ps1']
-
-    @ps1.setter
-    def ps1(self, value: str) -> None:
-        if not isinstance(value, str):
-            raise TypeError("hook.ps1: must be a string")
-        singletons['hook.ps1'] = value
-
-    @property
-    def ps2(self) -> str:
-        return singletons['hook.ps2']
-
-    @ps2.setter
-    def ps2(self, value: str) -> None:
-        if not isinstance(value, str):
-            raise TypeError("hook.ps2: must be a string")
-        singletons['hook.ps2'] = value
-
 PysUndefined()
-PysHook()